pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "socialgene"
version = "0.1.11"
description = "Creating and interacting with graph databases of protein domains and their genome coordinates"
readme = "README.md"
authors =  [
    {name = "Chase M. Clark", email = "chasingmicrobes@gmail.com"},
]
keywords = []  #! TODO
# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Unix",
  "Operating System :: POSIX",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
        "pandas>=1.3",
        "numpy>=1.21",
        "neo4j>=5.9.0",
        "biopython>=1.7",
        "textdistance>=4.2.1",
        "rich",
        "requests"
]

[project.license]
text = "MIT"

[project.urls]
repository = "https://github.com/socialgene/sgpy"
homepage = "https://socialgene.github.io"

[project.optional-dependencies]
ci = ["pytest-cov"]
cd = ["twine"]

[project.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
socialgene                     ="socialgene.cli.__main__:main"

# Used in the Nextflow pipeline
sg_process_domtblout    = "socialgene.cli.nextflow.process_domtblout:main"
sg_clean_hmm            = "socialgene.cli.nextflow.clean_hmms:main"
sg_process_genbank      = "socialgene.cli.nextflow.export_protein_loci_assembly_tables:main"
sg_ncbi_taxonomy        = "socialgene.cli.nextflow.parse_ncbi_taxonomy:main"
sg_export_parameters    = "socialgene.cli.nextflow.parameter_export:main"
sg_export_neo4j_headers = "socialgene.cli.nextflow.export_neo4j_header_files:main"
sg_create_neo4j_db      = "socialgene.cli.nextflow.create_neo4j_db:main"
sg_hmm_tsv_parser       = "socialgene.cli.nextflow.socialgene_hmm_tsv_parser:main"
sg_get_goterms          = "socialgene.utils.goterms:main"

# search
sg_mm_create            = "socialgene.mmseqs.create_database:main"
sg_mm_search            = "socialgene.mmseqs.search:main"
# Modify database
sgdb_import_classyfire = "socialgene.dbmodifiers.classyfire.import:main"

# neo4j
sg_indices              = "socialgene.cli.neo4j.indices:main"
sg_massage              = "socialgene.cli.neo4j.massage:main"

# info
sg_version              = "socialgene.utils.version:main"
sg_neo4j_version        = "socialgene.utils.version:neo4j"
sg_nodes                = "socialgene.neo4j.schema.nodes:printer"
sg_relationships        = "socialgene.neo4j.schema.relationships:printer"
